Upgrade puppeteer to 19.2.2 and fix date problem
Renderer stopped working on the 2nd of November 2022 because
there is a month change between Monday 31st and Wednesday 2nd, which
was causing a problem when iterating through several days of the week
to find one with data available. Monday 31st and Tuesday 1st didn't have
any table available on the webpage, and a weird Date.setDate() call was
being done with values 33, 34... (currentDate.getDate() + offset). This was
returning weird dates in the future and no container was being found.